[Develop] Move SharedStorageType Enum from cluster_config.py to common.py
